+++ span_marker/trainer.py
@@ -1,9 +1,11 @@
 import logging
-from typing import Callable, Dict, List, Optional, Tuple
+import math
+from typing import Any, Callable, Dict, List, Optional, Tuple
 
 import torch
 from datasets import Dataset
 from torch.utils.data import DataLoader
+from tqdm.autonotebook import tqdm
 from transformers import (
     EvalPrediction,
     TrainerCallback,
@@ -32,9 +34,11 @@ class Trainer(TransformersTrainer):
             The arguments to tweak for training. Will default to a basic instance of :class:`~transformers.TrainingArguments` with the
             ``output_dir`` set to a directory named *models/my_span_marker_model* in the current directory if not provided.
         train_dataset (Optional[~datasets.Dataset]):
-            The dataset to use for training.
+            The dataset to use for training. Must contain ``tokens`` and ``ner_tags`` columns, and may contain
+            ``document_id`` and ``sentence_id`` columns for document-level context during training.
         eval_dataset (Optional[~datasets.Dataset]):
-             The dataset to use for evaluation.
+            The dataset to use for evaluation. Must contain ``tokens`` and ``ner_tags`` columns, and may contain
+            ``document_id`` and ``sentence_id`` columns for document-level context during evaluation.
         model_init (Optional[Callable[[], SpanMarkerModel]]):
             A function that instantiates the model to be used. If provided, each call to :meth:`Trainer.train` will start
             from a new instance of the model as given by this function.
@@ -75,6 +79,9 @@ class Trainer(TransformersTrainer):
         - **is_in_train** -- Whether or not a model is currently running :meth:`~Trainer.train` (e.g. when ``evaluate`` is called while
           in ``train``)
     """
+
+    REQUIRED_COLUMNS: Tuple[str] = ("tokens", "ner_tags")
+    OPTIONAL_COLUMNS: Tuple[str] = ("document_id", "sentence_id")
 
     def __init__(
         self,
@@ -155,10 +162,14 @@ class Trainer(TransformersTrainer):
         Returns:
             Dataset: The normalized and tokenized version of the input dataset.
         """
-        for column in ("tokens", "ner_tags"):
+        for column in self.REQUIRED_COLUMNS:
             if column not in dataset.column_names:
                 raise ValueError(f"The {dataset_name} dataset must contain a {column!r} column.")
 
+        # Drop all unused columns, only keep "tokens", "ner_tags", "document_id", "sentence_id"
+        dataset = dataset.remove_columns(
+            set(dataset.column_names) - set(self.OPTIONAL_COLUMNS) - set(self.REQUIRED_COLUMNS)
+        )
         # Normalize the labels to a common format (list of label-start-end tuples)
         dataset = dataset.map(
             label_normalizer,
@@ -168,12 +179,159 @@ class Trainer(TransformersTrainer):
         # Tokenize and add start/end markers
         with tokenizer.entity_tracker(split=dataset_name):
             dataset = dataset.map(
-                lambda batch: tokenizer(batch["tokens"], labels=batch["ner_tags"], return_num_words=is_evaluate),
+                lambda batch: tokenizer(batch, return_num_words=is_evaluate),
                 batched=True,
-                remove_columns=dataset.column_names,
+                remove_columns=set(dataset.column_names) - set(self.OPTIONAL_COLUMNS),
                 desc=f"Tokenizing the {dataset_name} dataset",
             )
+        # If "document_id" AND "sentence_id" exist in the training dataset
+        if {"document_id", "sentence_id"} <= set(dataset.column_names):
+            dataset = dataset.sort(column_names=["document_id", "sentence_id"])
+            # TODO: Sort the dataset
+            dataset = self.add_context(
+                dataset,
+                tokenizer.model_max_length,
+                max_prev_context=self.model.config.max_prev_context,
+                max_next_context=self.model.config.max_next_context,
+            )
+        # Spread between multiple samples where needed
+        original_length = len(dataset)
+        dataset = dataset.map(
+            lambda sample: Trainer.spread_sample(
+                tokenizer.model_max_length, self.model.config.marker_max_length, sample
+            ),
+            batched=True,
+            batch_size=1,
+            desc="Spreading data between multiple samples",
+        )
+        new_length = len(dataset)
+        logger.info(
+            f"Spread {original_length} sentences across {new_length} samples, "
+            f"a {(new_length / original_length) - 1:%} increase. You can increase "
+            "`model_max_length` or `marker_max_length` to decrease the number of samples, "
+            "but recognize that longer samples are slower."
+        )
         return dataset
+
+    @staticmethod
+    def add_context(
+        dataset: Dataset,
+        model_max_length: int,
+        max_prev_context: Optional[int] = None,
+        max_next_context: Optional[int] = None,
+    ) -> Dataset:
+        """Add document-level context from previous and next sentences in the same document.
+
+        Args:
+            dataset (`Dataset`): The partially processed dataset, containing `"input_ids"`, `"start_position_ids"`,
+                `"end_position_ids"`, `"document_id"` and `"sentence_id"` columns.
+            model_max_length (`int`): The total number of tokens that can be processed before
+                truncation.
+            max_prev_context (`Optional[int]`): The maximum number of previous sentences to include. Defaults to None,
+                representing as many previous sentences as fits.
+            max_next_context (`Optional[int]`): The maximum number of next sentences to include. Defaults to None,
+                representing as many previous sentences as fits.
+
+        Returns:
+            Dataset: A copy of the Dataset with additional previous and next sentences added to input_ids.
+        """
+        all_input_ids = []
+        all_start_position_ids = []
+        all_end_position_ids = []
+        for sample_idx, sample in tqdm(
+            enumerate(dataset), desc="Adding document-level context", total=len(dataset), leave=False
+        ):
+            # Sequentially add next context, previous context, next context, previous context, etc. until
+            # max token length or max_prev/next_context
+            tokens = sample["input_ids"][1:-1]
+            start_position_ids = sample["start_position_ids"]
+            end_position_ids = sample["end_position_ids"]
+
+            next_context_added = 0
+            prev_context_added = 0
+            remaining_space = model_max_length - len(tokens) - 2
+            while remaining_space > 0:
+                next_context_index = sample_idx + next_context_added + 1
+                should_add_next = (
+                    (max_next_context is None or next_context_added < max_next_context)
+                    and next_context_index < len(dataset)
+                    and dataset[next_context_index]["document_id"] == sample["document_id"]
+                )
+                if should_add_next:
+                    # TODO: [1:-1][:remaining_space] is not efficient
+                    tokens += dataset[next_context_index]["input_ids"][1:-1][:remaining_space]
+                    next_context_added += 1
+
+                remaining_space = model_max_length - len(tokens) - 2
+                if remaining_space <= 0:
+                    break
+
+                prev_context_index = sample_idx - prev_context_added - 1
+                should_add_prev = (
+                    (max_prev_context is None or prev_context_added < max_prev_context)
+                    and prev_context_index >= 0
+                    and dataset[prev_context_index]["document_id"] == sample["document_id"]
+                )
+                if should_add_prev:
+                    # TODO: [1:-1][remaining_space:] is not efficient
+                    prepended_tokens = dataset[prev_context_index]["input_ids"][1:-1][-remaining_space:]
+                    tokens = prepended_tokens + tokens
+                    # TODO: Use numpy? np.array(sample["start_position_ids"]) + len(prepended_tokens)
+                    start_position_ids = [index + len(prepended_tokens) for index in start_position_ids]
+                    end_position_ids = [index + len(prepended_tokens) for index in end_position_ids]
+                    prev_context_added += 1
+
+                if not should_add_next and not should_add_prev:
+                    break
+
+                remaining_space = model_max_length - len(tokens) - 2
+
+            all_input_ids.append([sample["input_ids"][0]] + tokens + [sample["input_ids"][-1]])
+            all_start_position_ids.append(start_position_ids)
+            all_end_position_ids.append(end_position_ids)
+
+        dataset = dataset.remove_columns(("input_ids", "start_position_ids", "end_position_ids"))
+        dataset = dataset.add_column("input_ids", all_input_ids)
+        dataset = dataset.add_column("start_position_ids", all_start_position_ids)
+        dataset = dataset.add_column("end_position_ids", all_end_position_ids)
+
+        return dataset
+
+    @staticmethod
+    def spread_sample(
+        model_max_length: int, marker_max_length: int, sample: Dict[str, List[Any]]
+    ) -> Dict[str, List[Any]]:
+        """Spread sentences between multiple samples if lack of space per sample requires it.
+
+        Args:
+            model_max_length (`int`): The total number of tokens that can be processed before
+                truncation.
+            marker_max_length (`int`): The maximum length for each of the span markers. A value of 128
+                means that each training and inferencing sample contains a maximum of 128 start markers
+                and 128 end markers, for a total of 256 markers per sample.
+            sample (`Dict[str, List[Any]]`): A dictionary of dataset keys to singleton lists
+                of values.
+
+        Returns:
+            Dict[str, List[Any]]: A dictionary of dataset keys to not strictly singleton lists
+                of values.
+        """
+        sample = {key: value[0] for key, value in sample.items()}
+        total_sample_length = model_max_length + 2 * marker_max_length
+        sample_marker_space = (total_sample_length - len(sample["input_ids"])) // 2
+        spread_between_n = math.ceil(len(sample["start_position_ids"]) / sample_marker_space)
+        spread_samples = []
+        for i in range(spread_between_n):
+            sample_copy = sample.copy()
+            start = i * sample_marker_space
+            end = (i + 1) * sample_marker_space
+            sample_copy["start_position_ids"] = sample["start_position_ids"][start:end]
+            sample_copy["end_position_ids"] = sample["end_position_ids"][start:end]
+            if "labels" in sample:
+                sample_copy["labels"] = sample["labels"][start:end]
+            sample_copy["num_spans"] = len(sample_copy["start_position_ids"])
+            spread_samples.append(sample_copy)
+        return {key: [one_sample[key] for one_sample in spread_samples] for key in sample.keys()}
 
     def get_train_dataloader(self) -> DataLoader:
         """Return the preprocessed training DataLoader."""